requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.4.2
python-jose==3.3.0
uuid==1.30
python-multipart==0.0.6
starlette==0.27.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.0.0
requests==2.31.0
redis==5.0.1
slowapi==0.1.8
orjson==3.9.10
